# feat(turbo): implement smart pipeline with selective task execution (MAR-41)