--- a/apps/backend-api/package.json
+++ b/apps/backend-api/package.json
@@ -47,6 +47,7 @@
     "pnpm": ">=10.0.0"
   },
   "dependencies": {
+    "@ai-fastify-template/config": "workspace:*",
     "@fastify/autoload": "^6.0.0",
     "@fastify/sensible": "^6.0.0",
     "fastify": "^5.0.0",