--- a/apps/backend-api/src/server.ts
+++ b/apps/backend-api/src/server.ts
@@ -1,22 +1,30 @@
+import { getPort, getHost, isDevelopment } from '@ai-fastify-template/config';
 import Fastify from 'fastify';
 
 import app from './app.js';
 
 const server = Fastify({
-  logger: {
-    level: 'info',
-    transport: {
-      target: 'pino-pretty',
-    },
-  },
+  logger: isDevelopment()
+    ? {
+        level: 'info',
+        transport: {
+          target: 'pino-pretty',
+        },
+      }
+    : {
+        level: 'info',
+      },
 });
 
 server.register(app);
 
 const start = async (): Promise<void> => {
   try {
-    await server.listen({ port: 3000, host: '0.0.0.0' });
-    server.log.info('Server listening on http://localhost:3000');
+    const port = getPort();
+    const host = getHost();
+
+    await server.listen({ port, host });
+    server.log.info(`Server listening on http://${host}:${port}`);
   } catch (err) {
     server.log.error(err);
     throw new Error('Failed to start server');