--- a/packages/config/.prettierignore
+++ b/packages/config/.prettierignore
@@ -1,0 +1,2 @@
+build/
+coverage/