--- a/packages/config/src/index.ts
+++ b/packages/config/src/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Configuration utilities for the AI Fastify Template
+ */
+
+export * from './server-config.js';
+export * from './environment.js';