@mkwatson 2025-06-23
## 🔧 CI Fix Applied - PR Now Ready for Merge

### **Issue Resolved**
Fixed CI failure: `Cannot find module '@ai-fastify-template/config'` during TypeScript compilation.

### **Root Cause** 
The `type-check` task in `turbo.json` was running before workspace dependencies were built, causing module resolution to fail in clean CI environments.

### **Solution Applied**
Updated `turbo.json` to add proper dependency order:

```diff
"type-check": {
- "dependsOn": ["lint"],
+ "dependsOn": ["^build", "lint"],
  "outputs": []
}
```

### **Why This Works**
- `^build` ensures all workspace packages are built before type-checking
- Follows standard Turbo patterns for monorepo dependency management  
- Local development now matches CI behavior exactly
- No custom scripts needed - pure tool-first approach

### **Verification**
✅ **CI now passes** - All quality gates successful  
✅ **Clean state testing** - Works from `rm -rf build && pnpm type-check`  
✅ **Local development** - No workflow changes needed  
✅ **Future-proof** - Prevents similar dependency issues

The PR is now ready for merge with robust CI/CD pipeline and comprehensive code review feedback addressed.

🚀 Generated with [Claude Code](https://claude.ai/code)